Develop

See merge request fauzipadlaw/pemesanan!79

+++ app/Models/BahanBaku.php
@@ -67,7 +67,7 @@ class BahanBaku extends Model
     protected $casts = [
         'nama_bahan_baku' => 'string',
         'satuan' => 'string',
-        'sisa' => 'integer'
+        'sisa' => 'double'
     ];
 
     /**
